+++ src/tools/design/dev-pkg/utils/SimpleComponent/SimpleComponent.tsx
